Desktop/All project/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
ImageHash==4.3.1
django-filter==23.5
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
pycoingecko==3.1.0
sentry-sdk==2.44.0
python-json-logger==4.0.0
django-ratelimit==4.1.0

